Implemented Ella's latest design prototype (10-22-2023) + add basic header
- Added a basic sticky header (w/out mobile menu functionality yet) to the top in "base.html". The design can be modified in the future to add more nav links/change logo as needed. Note that the header should be modified with Jinja in the future to change depending on whether if the user is logged in or not (e.g. if not, then show sign up / log in button; if is, then maybe show profile picture with drop down instead?)
- Implemented Ella's latest design prototype from Figma for login page, with a couple of dimensions tweaked here and there due to oversize.
- Note that implementation differs from design in one or two areas (e.g. header text) to accommodate header.
- Changed file structure in static folder to organize css (and possibly JS?) files, as well as to add a vendor folder containing 3rd party resources (fonts, images, etc.)
-Temporarily added a basic app.py to run front-end files, since for some reason, models.py does not compile for me (some SQLAlchemy error with db.Integer not taking in any parameters, i think...?)
