+++ app/src/app.py
@@ -1,14 +1,21 @@
-from flask import Flask, render_template
-from models import db, User
+from flask import Flask, render_template, redirect
+# from auth import auth
 
 app = Flask(__name__)
-
-# Configure and initalize database
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
-db.init_app(app)
+# app.register_blueprint(auth)
 
 @app.route('/')
 def index():
     return render_template('index.html')
 
-db.create_all()
+# Incorporated within auth.py blueprint
+@app.route('/login.html')
+def redirectLogin():
+    return redirect('/login', code = 302)   # Probably should change 302 redirect code later
+
+@app.route('/login')
+def login():
+    # Temp logged in value for now
+    loggedIn = False
+    return render_template('login.html', logged_in = loggedIn)
+
