--- app/src/static/base.css
@@ -1,1 +1,0 @@
-/* base.css - contains styling for base template of website (headers, footers, etc.) */
