--- app/src/static/login.css
@@ -1,1 +1,0 @@
-/* login.css - contains styling for the login page specifically, barring base elements */
